requirements.txt
fastapi>=0.95.0,<1.0.0
uvicorn[standard]>=0.22.0,<1.0.0
pydantic
python-multipart>=0.0.6,<1.0.0
passlib[bcrypt]>=1.7.0,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
sqlalchemy>=1.4.0,<2.0.0
psycopg2-binary>=2.9.0,<3.0.0
opencv-python>=4.7.0,<5.0.0
numpy>=1.24.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic-settings
